-Added better vjDebug output commands -Added ability to check app for dependencies -Fixed cubesApp to check dependencies
git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@1563 08b38cba-cd3b-11de-854e-f91c5b6e4272

+++ juggler/modules/vrjuggler/vrj/Draw/OGL/GlContextData.h
@@ -103,7 +103,7 @@ protected:
       int context_id = getCurContext();
 
       // Make sure that we will reference a valid element
-      while(mContextDataVector.size() <= context_id)
+      while((int)mContextDataVector.size() <= context_id)
       {
          mContextDataVector.push_back(ContextDataType());
          vjDEBUG(vjDBG_DRAW_MGR,3) << "Adding ContextDataVector element: size now: " << mContextDataVector.size() << endl << vjDEBUG_FLUSH;
